frontend/pdf/print: set type to module in package.json
And switch to esm syntax for the matchingTranslationKeys eslint plugin.

Because there cannot be dashes in an identifier, the rule is now called
'matchingTranslationKeys'.

jest-serializer-vue-tjw loads the vue config with require,
and seems to expect a commonjs module.
We switched to a esm module, and now the structure is not
pluginOptions.jestSerializer,
but
default.pluginOptions.jestSerializer
-> with the added export we can support vue itself and jest-serializer-vue-tjw.

This might enable us to update vite-plugin-comlink
- https://github.com/ecamp/ecamp3/pull/6220
And might help with
- https://github.com/ecamp/ecamp3/issues/6038

Fix checked in dist files in pdf/dist/