e2e: wait for the requests to either have cache 'HIT', 'MISS', 'PASS'
Because this is a distributed system, while a PATCH request is running,
other clients may still do a dirty read of the cached value to be invalidated.
Symfony also splits up the purge requests, so this may take a while.

As an observer from the point of the e2e tests (or any clients using our application),
it's possible (by design) to read the previous value, which is not valid anymore when
the PATCH, POST or DELETE request has ended.

Thus we don't need to assert this strictly in cypress, we can repeat the get
request for some time until we observe the desired behaviour.

This is better than retrying the whole test because the retry is very specific with
the request which should have a more or less strict defined caching behaviour.

Issue: #5493