Backtrack on using broken composition api
Before this commit, all "required" validations in all dialogs were
broken and would always show an error even when the fields were filled.
As far as I can tell this was because of a strange interplay between
vee-validate in the version we are using and reactive() or ref() of the
Vue reactivity system.

Before this commit, entityData (the local state defined in the base
dialog) was extracted into a composable useEntityData. All dialogs
inheriting from the base dialog would have a
v-model="entityData.someField" on their e-text-fields. Vee-validate
then was not aware of any contents filled into the text fields, and
would always treat the value as "" (empty string). Therefore, the
required validations would always fail and prevent the dialog forms from
being submitted.

The situation was also not made better due to our inheritance structure
(extends: DialogBase in all the dialogs), which is not really supported
when using the composition API. So parts of the base class had to be
re-implemented in some of the specific dialogs.

In this commit, we remove the useEntityData composable again and return
to the options API. The only reason we needed access to entityData in
the composition API was that we have to call one setter method when an
entity URL from the clipboard is resolved. So we cheat a little bit here
and gain access to the options API method through getCurrentInstance(),
which is undocumented but well known. This might break in the future,
but it's still way better than dealing with compatibility bugs in an
outdated version of vee-validate until we can finally upgrade to Vue 3.