print: increase --max-header-size for development
The http module of node has a max header size for the http server and the client.
When we send an http request to the api and get too many xkey entries,
the client request fails with: "Parse Error: Header overflow"
Both values are configured with the --max-http-header command line option.

Don't set this parameter in production, because there we don't emit
the XKey headers to the clients.
And there we want to see if we accidentally generate too large headers.