CI: parallelize docker build
And simplify adding new image.
Now you just have to add a docker compose with an image name and a build section.

Also unset the image of api for dev, some people had problems when
there was an image defined which cannot be downloaded.

Populate the buildArgValues in the build-and-push step
because passing secrets between jobs seems to be discouraged by github.
https://github.com/orgs/community/discussions/37942