refresh the access token in the background
Use gesdinet/jwt-refresh-token-bundle to achieve this.
The access token lifetime was reduced to 30 mins
according to https://cloud.google.com/apigee/docs/api-platform/antipatterns/oauth-long-expiration
as a starting point.
Do not log out if the refresh in the background fails, the user might want to backup
the content he is currently editing.

One proposed pattern is to intercept 401 responses and refresh the token when the requests fail.
Because we often have multiple requests running, this did not work well.
Now we have a timeout in the background which refreshes the access token periodically.

closes #4898