chore: rewrite ecamp3 deployment to helmfile
That it is easier to use github action secrets and vars in the deployment.
Leave dev and stage-prod in separate files that the review is somehow easier.
They will be put into one workflow in the next commit.

The default false for the num_threads is currently a hack because
the if statement in the secret should be an if empty, not just if.
There we should move the default into the application code, and
then we can simplify the null handling.

Also add api.dataMigrationsDir as an explicit value.
This was implicit before.
There we should also move the default to the application code later.

I leave these improvements out that this PR remains reviewable.
It's also best to diff the deleted values.yaml and the values.yaml.gotmpl.
Sadly the similarity was not enough for git to mark this as rename.