Reload schedule entries when changing the numbering style of an activity
This is a bugfix for the following scenario:
- Load the picasso
- Open the activity detail of an activity
- Change the activity's category to one with a different numbering style

Expected outcome: All displayed instances of the schedule entry show the
schedule entry number change, and all schedule entries at all times have
the correct up-to-date schedule entry number.

Actual outcome: The color of the edited schedule entry changes, but the
schedule entry number on the top left remains unchanged. Also, when
navigating back to the picasso, the schedule entry numbering of other
schedule entries is outdated.