api: increase default memory limit to 500M in development
With php 8.3 it got even cloud native and allows to set ini values from env vars:
https://www.php.net/manual/en/configuration.file.php#example-33
Can be overridden in .env.
If the variable is not set, the php default (128M) is used.

I did not set the default in the api-platform.dev.ini because
if I used the fallback empty string in the docker compose (that docker compose
does not complain that it is not defined), the fallback
in the php ini is not used.
Do not use api/.env to set the default value because we don't need that for prod.
Another valid place to set the value would be the development stage in the docker build.

Issue: #8036