Fix vite reloading the page constantly
Pre-optimize all dependencies which we know are present on the page.
It's way better for DX to optimize a little too much than too little and
have Vite reload the page and throwing you back to the start page when
navigating the app for the first time after starting the frontend.
Taken and adapted from https://github.com/vuetifyjs/vuetify-loader/issues/350#issuecomment-3129253379