frontend: fix ECheckbox.spec.js
Setting up vuetify is now simpler, it just needs the setupVuetify function at the start.
In setupVuetify we need the Vuetify plugin, not the vuetify instance.
It is now easier to make ad hoc components for tests, no Vue.component needed anymore.
You cannot use the wrapper for snapshots anymore, you need wrapper.html().

Vuetify internals: the checkbox always has the attribute value with the value "true".
The difference between checked and unchecked is the checked attribute.

I also did not find a way to fetch the current values of the data of a component.
That's why we have an additional span with the value of data.

I also could remove an unnecessary resetAllMocks.